--- metals/src/main/scala/scala/meta/internal/zipkin/ZipkinUrls.scala
@@ -1,7 +1,0 @@
-package scala.meta.internal.zipkin
-import scala.meta.internal.pantsbuild.commands.BloopGlobalSettings.Property
-
-object ZipkinUrls {
-
-  val zipkinServerUrl: Property = Property("metals.zipkin.server.url")
-}
